refactor(client): separate Agent and probes
Separate packages for:

- Probes (#5)
  - User Agent (#4)
  - Timezone offset (#6)
  - System fonts (#7)
