+++ client/packages/agent/src/Agent.js
@@ -1,13 +1,46 @@
-const probes = require('@pouk/idem-client-probes')
+const Probes = require('@pouk/idem-client-probes')
 
-function detect () {
-  console.log(Object.keys(probes))
+const serialize = require('./serialize')
+const encode = require('./encode')
 
-  return Promise.resolve({ id: 'xxx' })
+function Agent () {
+  // use w/o `new` keyword
+  if (!(this instanceof Agent)) {
+    return new Agent()
+  }
+
+  const probes = {}
+
+  for (const name in Probes) {
+    const make = Probes[name]
+    probes[name] = make()
+  }
+
+  this.probes = probes
+
+  return this
 }
 
-module.exports = function () {
+Agent.prototype.detect = async function () {
+  const { probes } = this
+
+  const data = {}
+
+  for (const key in probes) {
+    const probe = probes[key]
+
+    data[key] = await probe()
+  }
+
+  const id = encode(serialize(data))
+
   return {
-    detect
+    id,
+    data
   }
 }
+
+module.exports = Agent
+
+module.exports.encode = encode
+module.exports.serialize = serialize
