+++ client/packages/agent/src/encode.js
