+++ client/packages/agent/src/serialize.js
