+++ client/packages/agent/test/Agent.spec.js
@@ -1,0 +1,8 @@
+import test from 'ava'
+
+import Agent from '..'
+
+test('constructor', t => {
+  t.true(new Agent() instanceof Agent)
+  t.true(Agent() instanceof Agent)
+})
