+++ client/packages/probe-system-fonts/package.json
@@ -15,8 +15,14 @@
     "type": "git",
     "url": "git+https://github.com/pouk/idem.git"
   },
-  "scripts": {},
+  "scripts": {
+    "test": "ava",
+    "test:converage": "nyc ava -v"
+  },
   "bugs": {
     "url": "https://github.com/pouk/idem/issues"
+  },
+  "dependencies": {
+    "font-detect": "^1.0.0"
   }
 }
