--- client/packages/probe-timezone-offset/src/TimezoneOffset
@@ -1,3 +1,0 @@
-module.exports = function () {
-  return () => Promise.resolve('xxx')
-}
