+++ client/packages/probe-timezone-offset/test/factory.spec.js
@@ -1,0 +1,8 @@
+import test from 'ava'
+
+import factory from '..'
+
+test('signature', t => {
+  t.is(typeof factory, 'function')
+  t.is(typeof factory(), 'function')
+})
