+++ client/packages/probe-user-agent/src/UserAgent.js
@@ -1,3 +1,17 @@
-module.exports = function () {
-  return () => Promise.resolve('xxx')
+/**
+ * Factory for probe to get User Agent
+ *
+ * @returns {Function}
+ */
+
+const factory = () => {
+  function UserAgent () {
+    return window.navigator.userAgent
+  }
+
+  return UserAgent
 }
+
+// expose factory
+
+module.exports = factory
