+++ client/packages/probes/index.js
@@ -1,3 +1,3 @@
-module.exports.UserAgent = require('@pouk/idem-probe-user-agent')
-module.exports.TimezoneOffset = require('@pouk/idem-probe-timezone-offset')
-module.exports.SystemFonts = require('@pouk/idem-probe-system-fonts')
+module.exports.UserAgent = require('@pouk/idem-client-probe-user-agent')
+module.exports.TimezoneOffset = require('@pouk/idem-client-probe-timezone-offset')
+module.exports.SystemFonts = require('@pouk/idem-client-probe-system-fonts')
