--- client/src/Probes/index.js
@@ -1,3 +1,0 @@
-module.exports.UserAgent = require('./UserAgent')
-module.exports.TimezoneOffset = require('./TimezoneOffset')
-module.exports.FontList = require('./FontList')
