--- client/src/index.js
@@ -1,2 +1,0 @@
-module.exports.Probes = require('./Probes')
-module.exports.Agent = require('./Agent')
